updated dev files
