--- index.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<link rel="stylesheet" href="sudoku.css">
-<body>
-    <h1>Sudoku Game</h1>
-    <!-- <div id="sudoku-container">
-        <div id="score-container">
-            <div id="sudoku-header">Score: <span id="score">0</span></div>
-        </div>
-        <div id="time-container">
-            <div id="sudoku-header">Time: <span id="time">00:00</span></div>
-        </div>
-    </div> -->
-    <div id="sudoku-board"></div>
-    <div id ="sudoku-numbers" class = "number-row"></div>
-    <button id="new-game-button">New Game</button>
-    <button id="reset-game-button">Reset Board</button>
-    <button id="check-board-button">Check Board</button>
-</body>
-<script src="sudoku.js"></script>
-</html>
