+++ package.json
@@ -1,0 +1,8 @@
+{
+  "dependencies": {
+    "csv-parser": "^3.0.0",
+    "express": "^4.21.0",
+    "nodemon": "^3.1.4",
+    "sqlite3": "^5.1.7"
+  }
+}
