+++ public/index.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="sudoku.css">
+    <title>Sudoku Game</title>
+</head>
+<body>
+    <h1>Sudoku Game</h1>
+    <div id="sudoku-board"></div>
+    <div id="sudoku-numbers" class="number-row"></div>
+    <button id="new-game-button">New Game</button>
+    <button id="reset-game-button">Reset Board</button>
+    <button id="check-board-button">Check Board</button>
+    <script src="sudoku.js"></script>
+</body>
+</html>
