--- sudoku.css
@@ -1,155 +1,0 @@
-/* styles.css */
-body {
-    font-family: Arial, sans-serif;
-    text-align: center;
-    margin-top: 10px;
-    color:#48686f;
-}
-
-/* 
-#sudoku-container {
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 20px;
-}
-
-#score-container, #time-container {
-    font-size: 20px;
-    font-weight: bold;
-}
-
-#sudoku-header {
-    margin: 0;
-}
-
-#score, #time {
-    color: #333;
-    margin-left: 5px;
-} */
-
-
-#sudoku-board {
-    display: grid;
-    grid-template-columns: repeat(9, 40px);
-    grid-template-rows: repeat(9, 40px);
-    gap: 0;
-    margin: 0 auto;
-    width: fit-content;
-    border-left: 2px solid gray;
-    border-top: 2px solid gray;
-}
-
-
-#sudoku-numbers {
-    display: flex;
-    justify-content: center;
-    margin-top: 20px;
-}
-
-#sudoku-numbers .cell-used{
-    color: #b3e1ed;
-    background-color: #78a3ad;
-    cursor:pointer;
-}
-
-#sudoku-numbers .cell-used:hover{
-    color: #b3e1ed;
-    background-color: #78a3ad;
-    cursor:pointer;
-}
-
-#sudoku-numbers .selected {
-    color: white;
-    background-color: #48686f;
-}
-
-
-.cell-disabled {
-    background-color: #e0f1f5; 
-}
-
-.invalid-move{
-    color:red;
-}
-
-.sudoku-cell.highlight{
-    background-color: #ffecfd;
-}
-
-.sudoku-cell.num-highlight{
-    background-color: #fac8f3;
-}
-
-#sudoku-numbers div {
-    width: 40px;
-    height: 40px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 18px;
-    border: 1px solid #9b9797;
-    box-sizing: border-box;
-    color: black;
-    background-color: #e0f1f5;
-    cursor: pointer; 
-}
-
-#sudoku-numbers div:hover{
-    color:white; 
-    background-color: #48686f;
-}
-/* board */
-#sudoku-board input {
-    width: 40px;
-    height: 40px;
-    display: flex;
-    align-items: center;
-    text-align: center;
-    justify-content: center;
-    font-size: 18px;
-    border: 1px solid #bbb;
-    box-sizing: border-box;
-    cursor: pointer; 
-}
-
-#sudoku-board input:hover{
-    color: black;
-    background-color: #f2ccec;
-}
-
-#sudoku-board input:focus {
-    outline: none;
-    background-color: #f2ccec;
-}
-
-button {
-    margin-top: 10px;
-    padding: 10px 20px;
-    font-size: 16px;
-    background-color: #e0f1f5;
-    border-radius: 5px;
-    margin-bottom: 10px;
-    border: white;
-}
-
-button:hover{
-    background-color: #48686f;
-    color: white;
-    cursor: pointer;
-}
-
-
-#sudoku-board input:nth-child(3n) {
-    border-right: 2px solid gray;
-}
-#sudoku-board input:nth-child(27n+22),
-#sudoku-board input:nth-child(27n+23),
-#sudoku-board input:nth-child(27n+24),
-#sudoku-board input:nth-child(27n+25),
-#sudoku-board input:nth-child(27n+26),
-#sudoku-board input:nth-child(27n+27),
-#sudoku-board input:nth-child(27n+19),
-#sudoku-board input:nth-child(27n+20),
-#sudoku-board input:nth-child(27n+21) {
-    border-bottom: 2px solid gray;
-}
